+++ src/chunk.h
@@ -2,7 +2,7 @@
 
 #include "value.h"
 
-#include <stdint.h>
+#include <cstdint>
 #include <vector>
 #include <utility>
 
@@ -30,7 +30,7 @@ class Chunk : std::vector<Byte> {
 
     void write(Byte byte, int line);
     void write(Op op, int line);
-    int addConstant(Value value);
+    uint8_t addConstant(Value value);
     void disassemble(const char* name) const;
     int disassembleInstruction(int offset) const;
     Value getConstant(int idx) const;
